requirements.txt
Flask 
gspread 
pandas 
google-auth-oauthlib 
flask-login 
flask-wtf 
Levenshtein 
python-dateutil 
gunicorn
google-generativeai
flask-socketio
python-dotenv